refactoring and adding Master COntrol and Reponse Routing to Opt Problem
