+++ applications/OptimizationApplication/python_scripts/algorithms/algorithm_gradient_projection.py
@@ -14,6 +14,7 @@ from KratosMultiphysics.OptimizationApplication.utilities.logger_utilities impor
 from KratosMultiphysics.OptimizationApplication.utilities.logger_utilities import OptimizationAlgorithmTimeLogger
 from KratosMultiphysics.OptimizationApplication.utilities.list_collective_expression_utilities import CollectiveListCollectiveProduct
 from KratosMultiphysics.OptimizationApplication.utilities.list_collective_expression_utilities import CollectiveListVectorProduct
+from KratosMultiphysics.OptimizationApplication.utilities.optimization_problem_utilities import OutputGradientFields
 
 def Factory(model: Kratos.Model, parameters: Kratos.Parameters, optimization_problem: OptimizationProblem):
     return AlgorithmGradientProjection(model, parameters, optimization_problem)
@@ -48,6 +49,7 @@ class AlgorithmGradientProjection(Algorithm):
         parameters.ValidateAndAssignDefaults(self.GetDefaultParameters())
 
         self.master_control = MasterControl() # Need to fill it with controls
+        self._optimization_problem.AddComponent(self.master_control)
 
         for control_name in parameters["controls"].GetStringArray():
             control = optimization_problem.GetControl(control_name)
@@ -65,9 +67,11 @@ class AlgorithmGradientProjection(Algorithm):
         self.__line_search_method = CreateLineSearch(settings["line_search"], self._optimization_problem)
 
         self.__objective = StandardizedObjective(parameters["objective"], self.master_control, self._optimization_problem)
+        self._optimization_problem.AddComponent(self.__objective)
         self.__constraints_list: 'list[StandardizedConstraint]' = []
         for constraint_param in parameters["constraints"].values():
             constraint = StandardizedConstraint(constraint_param, self.master_control, self._optimization_problem)
+            self._optimization_problem.AddComponent(constraint)
             self.__constraints_list.append(constraint)
         self.__control_field = None
         self.__obj_val = None
@@ -163,9 +167,9 @@ class AlgorithmGradientProjection(Algorithm):
     @time_decorator()
     def Output(self) -> KratosOA.CollectiveExpression:
         self.algorithm_data.GetBufferedData()["control_field"] = self.__control_field.Clone()
-        self.__objective.OutputGradientFields(self._optimization_problem, True)
+        OutputGradientFields(self.__objective, self._optimization_problem, True)
         for constraint in self.__constraints_list:
-            constraint.OutputGradientFields(self._optimization_problem, constraint.GetStandardizedValue() > 0.0)
+            OutputGradientFields(constraint, self._optimization_problem, True)
         for process in self._optimization_problem.GetListOfProcesses("output_processes"):
             if process.IsOutputStep():
                 process.PrintOutput()
