+++ applications/OptimizationApplication/python_scripts/controls/master_control.py
@@ -192,3 +192,6 @@ class MasterControl:
 
     def Finalize(self) -> None:
         CallOnAll(self.__list_of_controls, Control.Finalize)
+
+    def GetName(self) -> str:
+        return "master_control"
